Fix most of Microsoft warnings

+++ src/file_system/lomse_file_system.cpp
@@ -1,6 +1,6 @@
 //---------------------------------------------------------------------------------------
 // This file is part of the Lomse library.
-// Lomse is copyrighted work (c) 2010-2016. All rights reserved.
+// Lomse is copyrighted work (c) 2010-2018. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without modification,
 // are permitted provided that the following conditions are met:
@@ -282,7 +282,7 @@ long LocalInputStream::read (unsigned char* pDestBuffer, long nBytesToRead)
     //requested number of bites if the end of stream is reached.
 
     m_file.read((char*)pDestBuffer, nBytesToRead);
-    return m_file.gcount();
+    return long(m_file.gcount());
 }
